Fix and refactor insert-in-between actions
