+++ packages/admin/cms-admin/src/blocks/common/blockRow/BlockRow.sc.ts
@@ -28,7 +28,7 @@ const slideInAnimation = keyframes`
 export const BlockWrapper = styled("div")`
     position: relative;
 
-    :hover ${AddContainer} {
+    &:hover ${`${AddContainer}`} {
         display: block;
     }
 `;
